src/requirements.txt
discord.py==2.5.2
Flask==3.1.0
pywin32==310
winshell==0.6
yt_dlp==2025.3.27
PyNaCl==1.5.0